+++ lib/domain/i_entity.dart
@@ -5,12 +5,6 @@
 
 import 'package:equatable/equatable.dart';
 
-/// 值对象都要实现本接口
-abstract class IVerifiable {
-  // 判断输入值是否合法
-  bool isValid();
-}
-
 /// 实体类实现本接口
 abstract class IEntity extends Equatable {
   String get id;