+++ lib/get_arch_core.dart
@@ -3,15 +3,15 @@ library get_arch_core;
 export 'application/i_usecase.dart';
 
 export 'domain/i_failure.dart';
-export 'domain/i_common.dart';
+export 'domain/i_entity.dart';
 export 'domain/extension.dart';
 export 'package:dartz/dartz.dart' show Either;
 export 'package:rxdart/rxdart.dart';
 
 export 'interface/i_dto.dart';
 
+export 'profile/i_get_arch_package.dart';
 export 'profile/get_arch_application.dart';
 
 export 'package:get_it/get_it.dart';
-export 'package:injectable/injectable.dart' hide test;
 export 'package:equatable/equatable.dart';